# compressor/zstd: improvements