- static linking #define to get the advanced stuff
- hint src size to compressor
- no flush needed before end
- check for error

It's about twice as 25% faster now for 16KB chunks.

Note that a quick test to use resetCStream and reusing the same stream
did not appear to have any effect.

Signed-off-by: Sage Weil <sage@redhat.com>