@Cyan4973 2017-11-10
This is indeed below performance expectation.
I should have a look into it to understand why.
The first step would be to reproduce those benchmark numbers.