@liewegas 2017-11-10
See https://github.com/ceph/ceph/pull/18881.  Thanks for taking a look!  If we're lucky it's something silly like missing optimization.