@liewegas 2017-11-12
http://pulpito.ceph.com/sage-2017-11-12_17:15:27-rados-wip-sage2-testing-2017-11-12-0922-distro-basic-smithi/